lab_5b/pyproject.toml
[project]
name = "lab_5b"
version = "0.1.0"
description = "GRPO Finetuning Experiments (Qwen3-4B Reasoning & GPT-OSS-20B Code Optimization)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "comet_ml",
    "unsloth",
    "transformers",
    "datasets",
    "trl",
    "huggingface_hub",
    "python-dotenv",
    "vllm",
    "pandas",
    "numpy",
]

[tool.uv]
dev-dependencies = []
